Move 'Bundle list' component and styling to a separate folder.
