+++ client/my-sites/plans/jetpack-plans/product-store/bundle-list/style.scss
@@ -1,0 +1,25 @@
+.jetpack-product-store__bundles-list--featured-item {
+	display: flex;
+	flex-direction: column;
+}
+
+.jetpack-product-store__bundles-list--featured-item .foldable-card {
+	box-shadow: none;
+	width: 100%;
+	margin: 0;
+	border-bottom: 1px solid var( --studio-gray-5 );
+}
+
+.jetpack-product-store__bundles-list--featured-item .foldable-card.is-expanded {
+	margin: 0;
+}
+
+.jetpack-product-store__bundles-list--featured-item .foldable-card.is-expanded .foldable-card__content {
+	padding: 0;
+}
+
+.jetpack-product-store__bundles-list--featured-item .foldable-card__main {
+	font-size: $font-body;
+	line-height: 24px;
+	font-weight: 600;
+}
